pyproject.toml
[project]
name = "ask-pesu"
version = "0.1.0"
description = "A FastAPI application for question answering about PES University using LangChain and Google Gemini."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-qdrant>=0.2.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "secret_required: marks tests that require secrets or environment variables (e.g. GEMINI_API_KEY)."
]

[tool.ruff.lint]
select = [
    "E", # Pycodestyle errors
    "F", # Pyflakes errors
    "W", # Pycodestyle warnings
    "UP", # Pyupgrade rules
    "I", # Import related rules
    "C90", # Complexity rules
    "D", # Documentation rules
    "ANN", # Annotations rules
    "TYP", # Type checking rules
    "N", # Naming rules
    "Q000", # Quotation style rules
    "RET", # Return type rules
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
    "tests/"
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
